afmongodb, afsql: Fix a race condition in the queue handling.
On occassions where we received a message, woke up the worker thread,
then received two two others before the worker finished, we had a high
chance of triggering an assertion, because we ended up calling the
insert before the parallel_push stuff managed to wake up the thread
and clear the callbacks.

Now we explicitly clear parallel_push before popping the queue head,
for the simple reason that we do not need it at that point anymore.

Many thanks to Balazs Scheidler <bazsi@balabit.hu> for pointing out
what happens.

Cc: syslog-ng-stable@balabit.hu
Signed-off-by: Gergely Nagy <algernon@balabit.hu>