nvtable: implement proper limitation on NVTable size
NVTable has a maximum size of 65535 * 4, due to limitation of the
data structure (length is stored in 16 bits shifted by two). However
this limit was not properly boundary checked and oversize NVTable
got syslog-ng to abort with an assertion failure.

This patch changes syslog-ng behaviour so that reallocation problems are
properly reported.

Reported-By: Peter Eisenlohr
Cc: syslog-ng-stable@balabit.hu
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>