value-pairs: change the default scope to empty for the template-function option parser
When parsing value-pairs with the bison parser (e.g. mongodb destination),
the default scope is empty. On the other hand when using the value-pairs
command line parser (e.g. format-json), the default scope was rfc3164.

The latter was changed for consistency and in order to make it really
simple to use _only_ the specified nv pairs.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>