afsql: Database connection health check
Implemented a health check for the database connection after a query
fails in afsql_dd_insert_db() that addresses an issue where, if a
thread has a valid connection instance but has lost the actual
connection somehow, log messages can potentially be dropped without
an attempt to reconnect.

An error message is generated similar to the initial connection
error message if the connection is found to be down.

The value for a thread's failed_message_counter is not incremented
if the database connection is found to be down as the connection's
state is the probable cause of the failed query.

The placement allows any existing or subsequent uses of goto error
to take advantage of the health check.

Signed-off-by: Erik Maciejewski <erikm@ymail.com>