logqueue: Maximized the number of the messages removed from the queue. (fixes: #24383)
Sometimes because of a race condition it is possible that there are more
messages in the log queue than the configured limit (log_fifo_size()).

In this situation the the log_queue_fifo_move_input_unlocked tries to
remove more elements from the thread queue than it contains and
because of this, it produces an invalid linked list, causing heap corruptions
and other nastiness.

Signed-off-by: Attila Szalay <sasa@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>