jsonparser: New parser module for JSON parsing.
The new parser supports json-c only, it does not work with
json-glib. For this reason, the configure script was modified a
little, so that it uses --with-json only. When json-c is found or
selected, then both the parser and the template function will be
built. If json-glib is found or selected, then only the template
function will be available.

The summary after configure now lists the determined settings.

The new parser is - so far - fairly simple: it parses a JSON with
json-c (int, double and string values are supported so far, booleans,
arrays and sub-objects aren't yet), and sets the appropriate
name-value pair in the LogMessage.

Users can configure a prefix to prepend before each key.

Usage is as simple as:

source s_json { tcp(port(21514) flags(no-parse)); };
destination d_json {
  file("/tmp/test.json"
       template("$(format-json --scope dot-nv-pairs)\n"));
};
parser p_json { json-parser (prefix(".json.")); };
log { source(s_json); parser(p_json); destination(d_json); };

Signed-off-by: Gergely Nagy <algernon@balabit.hu>