logproto: Fix log_proto_file_writer_flush()'s partial construction.
If the log messages has different length and only a partial write
happen the log_proto_file_writer_flush there are a possible buffer
under/overflow happen.

The problematic part is in the calculation about the last written
byte of the last written message. The calculation is not just too
difficult to follow but use the wrong message length in it.
Because of this ther are buffer under/overflow may happen or
starting to read the message in wrong position, causing messing the
log.

Signed-off-by: SZALAY Attila <sasa@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>