LogQueue: added keep_on_reload() method
This patch introduces a new member function for LogQueue instances. Its
role is to indicate whether we prefer keeping that LogQueue instance
around when reloading syslog-ng. The default FIFO implementation allows
the core to free the queue if it has zero items.

If the implementation doesn't supply such method, we default to keeping the
queue around.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>