scratch-buffers: Scratch buffer implementation.
Scratch buffers are thread-local GString buffer stacks. The intended
usage is that whenever a piece of code needs a scratch buffer for some
quick task, it can acquire one via scratch_buffer_acquire(), and
whenever it is not needed anymore, it releases it back to the stack
with scratch_buffer_release().

Buffers are allocated on-demand, and they're freed up at the end of
main_loop_io_worker_job_stop().

Code that wishes to use scratch buffers need to acquire and release as
appropriate. These return a ScratchBuffer object, from which the
GString part can be extracted with sb_string().

Signed-off-by: Gergely Nagy <algernon@balabit.hu>