jsonparser: Convert to use scratch buffers.
Instead of holding on to a pair of GStrings in the LogJSONParser
structure, and using them in a non-thread safe manner, use the newly
introduced scratch buffers instead.

Also, we now create a new JSON tokener for each message - it's thread
safe that way, and it's cheap enough to not need caching.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>