resolve_sockaddr: Fix domain stripping on initial lookup.
When looking up a hostname, we were only setting the 'positive' bool
during lookup, so it was only set when the host already existed in the
cache. When it didn't, and we just put it there right after the failed
lookup, the bool remained false.

This resulted in the domain part not being stripped even when
use_fqdn(no) was set.

The fix implemented in this patch is dead simple: if we managed to
resolve a host, we set the bool to TRUE.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>