value-pairs: Rekeying, rethought.
Instead of having a global list of transformations, we now bind the
transformations to a glob: for each glob, every transformation will be
applied, the glob is only matched against the key once.

This allows us to do far more fine-grained rewritings.

The new syntax is as follows:

value-pairs(scope(dot-nv-pairs)
            key("HOST")
            rekey(".cee.*"
                  shift(5)
                  add-prefix("Events.")
                  replace("Events.some_field" "some_field")
            ))

Given a message that would look like this in JSON:

{".cee.p_prod_name":"syslog-ng", ".cee.p_prod_id":"1234",
 ".cee.some_field":"This is cool!", ".cee.something":"456",
 "HOST":"localhost"}

With the above transformation, we'd end up with:

{"Events.p_prod_name":"syslog-ng", "Events.p_prod_id":"1234",
 "some_field":"This is cool!", "Events.something":"456",
 "HOST":"localhost"}

The rekey() only applies to keys matching the ".cee.*" glob (that is,
it's not applied to HOST), and the transformations will all be applied
onto every key that matches that glob. The glob will not be re-checked
by the transformation functions.

So, from ".cee.p_prod_name", we first get to "p_prod_name", then
prefix it with "Events.", and we even try to replace
"Events.some_field", but there's no such string prefix this time, so
the last transformation does nothing.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>