value-pairs: Make key() accept globs.
Instead of being a wrapper around pair(), turn key() into a
glob-accepting include statement instead (with keeping the name).

After this patch, the behaviour of key() changes: it no longer
guarantees that a given key will appear in the final set, as excludes
still have a chance to override it.

key() and exclude() patterns are now matched in order, and the last
one wins. The only way to guarantee a key into a set, in spite of any
exclude() rules, is via pair(). The key-value pairs specified by
pair() are still inserted into the set afterwards.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>