value-pairs: Fix a small memory leak in the commandline parser.
Whenever we parse a command-line, we construct a new argv array, to
prepend a zeroth element for glibs option parser. We correctly free'd
it on the error path, but never on the success path.

This lead to a small memory leak every time
value_pairs_new_from_cmdline() was called (that is, at the moment,
each $(format-json) template function leaked a little bit).

This patch fixes this issue by modifying the error handling a little:
instead of freeing everything and returning NULL, we only free the
value-pairs structure, and set it to NULL. The rest of the freeing is
left up to the common parts afterwards, which now correctly frees the
constructed argv array too.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>