gprocess: Fix --no-caps handling when building with hardening options.
When both glib and syslog-ng is built with hardening options enabled,
overwriting a string pointer with a FALSE boolean does not have the
expected results: the string will not become NULL.

In this case, --no-caps will end up with a segfault, because
process_opts.caps is non-NULL, and points to junk.

To fix this, introduce a new function that sets process_opts.caps to
NULL explicitly, and use this function as a callback, instead of
assuming that a FALSE boolean will have the expected results.

Reported-By: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>