configuration: allow configuration object duplicates
Historically syslog-ng used to allow overriding configuration block simply
by defining the same object twice. A previous patch has changed this
to an error, however there seem to be people who rely on the old behaviour.

In order to serve both groups of people (who get it wrong and who really
wants to use this) a new global variable has been introduced.

This line in the beginning of the config:

@define allow-config-dups 1

will restore the old behaviour, this is also printed to the console
at startup.

Reported-By: Dave Rawks <dave@pandora.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>