db-parser(): set .classifier.unknown tag if a message doesn't match
Whenever a message doesn't match the patterns, automatically set
the tag '.classifier.unknown'. Likewise if it does, clear this tag.

The rationale is to make it easier to use tags to detect if a message actually
matched while processing with the last db-parser().

We're clearing the same tag in case of a match so that if there are
multiple db-parser() instances, the check against .classifier.unknown
would work, even if the first one matched and the 2nd doesn't.

Reported-By: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>