logrewrite: Correctly initialize the conditional rules.
In case where the condition of a rewrite used a filter reference
instead of inlining the filter itself, the filter expression was never
initialized. This resulted in the condition always evaluating to
not-match, making the following construct completely useless:

 rewrite("new-value", value("MESSAGE"), condition(filter(f_my_filter)))

The fix is to introduce a new function, log_rewrite_init_method(),
which will initialize the condition filter if it exists, and has an
init method. Then we bind this method to LogPipe's init callback
during log_rewrite_init().

This way, when initializing the rewrite pipe, the condition will get
initialised too.

Reported-by: Thomas Wollner <tw@wollner-net.de>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>