[file_pos] Save the file positions when the message is written by the destination
Previous behaviour was that the file positions are saved after the message is queued, but
the problem was that if syslog-ng is crashed or stopped, but can't flush the queue at the end
The messages are lost which are in the queue, because the file position is updated,
but the message isn't written out.

This functionality required a lot of interface modifications to update positions in case of
message acknowledge phase.

Signed-off-by: Karsai Csaba <cskarsai@balabit.hu>