afprog: Terminate the child process on deinit.
When reloading the configuration, terminate the child process of a
program destination, for the following reasons: it will be restarted
anyway by afprog_dd_reopen(), so there's no harm done. But if dash is
used as /bin/sh, simply closing the stdin of the process will not make
it stop.

That, in turn, results in us starting the same program over and over
again on each SIGHUP. So instead of doing this, send a SIGTERM to the
child on deinit, too.

Reported-by: Thomas Wollner <tw@wollner-net.de>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>