afsql: Add support for specifying custom SQL backend options.
When one is using an SQL backend that supports a set of options that
syslog-ng does not set itself, yet, one does want these options set,
there was previously no way to accomplish that without changing the
source.

This patch adds a new option: dbd-option(OPTION_NAME VALUE) (where
OPTION_NAME is always a string, and value can be either a string, or a
number). The options specified this way will always be set with libdbi
whenever we (re)connect to the database.

A simple use-case:

destination d_slow_sql_test {
  sql(type("null")
      dbd-option("null.sleep.connect" 1)
      dbd-option("null.sleep.query" 5)
      columns("dummy")
      values("1")
  );
};

Signed-off-by: Gergely Nagy <algernon@balabit.hu>