afunix: clarified error messages in case of a failure
On Fedora systems, syslog-ng was configured to use unix-stream() /dev/log
whereas systemd supplied a unix-dgram() one, which caused difficult to
diagnose problems.

This patch adds further logging to this case and causes syslog-ng to fail
with an error message if it finds that /dev/log is using an incorrect
socket type.

Reported-By: Michal Schmidt <mschmidt@redhat.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>