dbparser: fixed message indexing in action rules with no correllation (incompatible)
Actions that are attached to rules having context-id (e.g. correllation),
use @0 for the message being generated and @1 the last message in
the context.

This is inconsistent with rules that have no correllation, where @0 was the
triggering message.

This patch fixes this behaviour, but at the same time it is incompatible
with previous versions.

Reported-By: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>