log_center_init_pipe_line(): rescope variables so most declarations are where the given variable is used
This patch is primarily meant to make the code a little bit more readable
by moving the variable declarations to where they are used instead of
being defined at the top of the function. junction {} handling will make
this function even more complex, so making it a bit more readable helps
a lot.

I also added a comment to describe the function so that it drives the
reader's eyes a bit for deeper understanding on what's going on.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>