afmongodb: Fix a race condition in the queue handling.
There is a race condition between afmongodb_dd_queue() and
afmongodb_worker_insert(), that can trigger an assertion when the rate
of incoming messages is very high.

The scenario that triggers this is like this:

* We receive a message, put it in our internal queue, and notify the
  worker thread.
* The worker thread resets the queue's parallel_push.
* We receive another message, notice that the queue is empty, so we
  insert it, and set parallel_push at the same time.
* The worker thread tries to pop the queue's head, but that triggers
  an assert, because parallel_push is set.

The root cause of the problem is that afmongodb_dd_queue() modifies
the queue without locking the queue_mutex. This patch corrects that
problem.

Reported-By: Costa Farber <costaf@wix.com>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>