afmongodb: Add support for safe-mode.
This adds a new option to the mongodb destination driver:
safe-mode(yes|no), which, as the name implies, can be used to turn
safe-mode on and off. For performance reasons, it is off by default.

Safe mode basically involves an extra check after each insert, that
checks whether the last insert succeeded, and only treating the insert
a complete success when the double check succeeds too.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>