configure.in: removed "environ" variable test in favor of an explicit OS dependency
The test didn't work reliably as "environ" as a variable is available on
Darwin in executable programs, but not in shared libraries. syslog-ng
uses that in a shared library, but there's no easy way to test that.

This patch changes the detection from an actual check of the environ
variable to a check to the underlying OS.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>