LogDestDriver: introduce log_dest_driver_queue_method()
This patch introduces a new derived queue method that has to be called
from all LogDestDriver instances. In this patch, the method is created
and all dest drivers explicitly call that in their _queue() method.

Note that in order to do this, destination drivers should pass a reference
and ack to this inherited method, which requires changes in all
destination drivers.

This change allows us to factor in more-and-more destination related
functionality in LogDestDriver instead of pushing those into
LogSourceGroup/LogDestGroup, which is going to be removed in a
follow-up patch. For example, previously group related counters were
updated in the LogSourceGroup/LogDestGroup objects, this is going to be
moved back to the driver, sparing us one additional LogPipe during
processing and making implementation a bit more object-oriented.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>