implement automatic loading of modules
Instead of having to rely on @module statements and --default-modules
parameters, this patch implements automatic loading of modules whenever
they are referenced from the configuration file.

It first enumerates all available modules, extracts the list of plugins
they offer and autoloads modules based on this information. This
means that it doesn't matter what the module name is, syslog-ng
will perfectly be aware which plugin is in which module.

Whenever multiple modules offer the same plugin (think about the
afsocket case) a "preference" value will determine which one is used.

The autoloading can be disabled by setting 'autoload-compiled-modules'
global variable set in the configuration file, which is now set to 1.

The SCL modules.conf file got removed too, instead of having to track
which modules need loading, the automatic means makes that unnecessary.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>