loggen: handle ENOBUFS gracefully
This patch adds handling of ENOBUFS by sleeping for 1msec one it occurs.
This fixes ENOBUFS handling on BSDs, which actually return this error
code even in blocking mode, thus causing loggen to exit with an error
when UDP is used and the network stack/interface is slower than
loggen.

Signed-off-by: Russ Milne <rtaitmilne@seccuris.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>