tfuuid: New template function to generate UUIDs.
This patch adds a new template function: $(uuid), which can be used to
generate Universally Unique IDentifiers.

The use case for this is to add an UUID to each message as soon as
they're received, so that they can be identified later, accross
destinations. This allows me to pre-generate an ID for messages going
into mongodb, and send those messages to a different destination too,
with the same ID included.

Use it like this:

rewrite r_add_uuid { set("$(uuid)" value("uuid")); };
log { source(s_network); rewrite(r_add_uuid);
      destination(d_mongo); destination(d_program); };

Signed-off-by: Gergely Nagy <algernon@balabit.hu>