value-pairs: Use a GPtrArray for explicit name=value pairs
Instead of storing the name-value pairs explicitly set by the user in
a GHashTable, store them in a GPtrArray. This has the advantage of
being both lighter and faster to both insert and traverse, at the cost
of making collissions more costy.

If there are two entries with the same key, the templates of both will
be evaluated, but only the latter will be used, which is a minor
performance hit. But not using GHashTable outweights that.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>