afmongodb: hack to link afmongodb properly against OpenSSL
In case the glib version is too old, afmongodb requires linking against
libcrypto, however even though libmongo-client.la lists it as a dependency
that is not used when linking modules with libtool.

This patch is a hack, a better solution would be welcome, it unconditionally
links the afmongodb plugin against libssl and libcrypto in case the internal
copy is being used.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>