logwriter: fixed a memory leak for messages that couldn't be flushed at once
Sometimes a message is dequeued, and then can't be posted to the destination
(probably because of an EAGAIN or something like that). In this case
it is readded back to the queue. In this case a reference was lost, causing
a memory leak, but not for all messages, but only for the affected ones.

Reported-By: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>