logtransport: handle ENOBUFS on output
FreeBSD returns ENOBUFS on send() failure instead of indicating
this conditions via poll().  The return of ENOBUFS actually is a send
error and is calulated in IP statistics, so the best is to handle it as
a success.  The only alternative would be to return EAGAIN, which could
cause syslog-ng to spin as long as buffer space is unavailable.  Since
I'm not sure how much time that would take and I think spinning the CPU
is not a good idea in general, I just drop the packet in this case.
UDP is lossy anyway

Reported-By: Traiano Welcome <Traiano.Welcome@mtnbusiness.co.za>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>