driver: fixed possible leak and use-after-free
log_dest_driver_release_queue() was possibly leaking the LogQueue
instances for file destinations when they got reaped. This was
caused by an earlier patch that fixed a crash in reloads,
more specifically this one: c7070e2a6f1c3a312260bcecf49d62028fef27ce

This patch should fix both cases properly, the leak in the file destination
driver and the original crash in the afsocket destination.

Also, this patch fixes a use-after-free condition, the next member of
a GList structure was referenced after it was removed from the list.

Kudos to Jakub for the detailed bug report and Algernon for the origianl
fix.

Reported-By: Jakub Jankowski <shasta@toxcorp.com>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>