debian: Fix the squeeze->wheezy install-remove-install upgrade
When one installed syslog-ng in squeeze, and removed it, then
dist-upgrades to wheezy, and tries to install syslog-ng again, it
won't automatically work, because the logrotate script was moved from
syslog-ng to syslog-ng-core, and the preinst wasn't clever enough to
deal with this situation.

The way to solve this is to move the .disabled file back to its place
under the following conditions (assuming the postinst is called with
'install' as its first parameter, and the .disabled file exists):

 * It's either an install when there was a former syslog-ng-core
   version.
 * Or it's an install where there was no former syslog-ng-core
   installation, and the destination file doesn't exist yet.

This supports both the install-remove-install case that span
squeeze->wheezy, and works even between wheezy->wheezy versions.

Reported-by: Andreas Beckmann <debian@abeckmann.de>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Closes: #665905