[PATCH (3.4) 2/3] afsql: Simplify insert_db()'s table validation
To make things a tiny bit easier, move the final error handling out of
_insert_db() into a separate function: afsql_dd_insert_fail_handler().

Furthermore, change afsql_dd_validate_table() to allocate the GString
for the table name itself, and resolve the template too, if so need
be.

The above two allows us to simplify afsql_dd_insert_db() a little:
when validating the table, we can immediately bail out, without a
goto, since we only need to free up the bare minimum.

And in the end, a monstrous if (success) {} else {} branch is replaced
by something a lot shorter, and clearer.

All of these together make the flow of afsql_dd_insert_db() easier to
understand and follow.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>