use on-demand LogMessage cloning to implement inmutable messages on parallel log paths
When processing a message, syslog-ng tries hard to avoid copying messages
along the line. The semantics provided by the syslog-ng configuration
is that parallel branches of the log processing pipeline are independent,
e.g. change on one of the branches may not affect the other one.

Earlier, this was implemented by the LogMultiplexer object, before
posting a message to a branch it checked if the branch may change
the message by inspecting the PIF_CLONE flag of the head element. This
didn't handle all cases (e.g. when a rewrite was applied between two
destinations).

This patch delegates this clone-when-needed operation right before the
actual change happens, this also allows some performance improvements
(since the clone may not always be needed).

I introduced the notion of "write-protection" to LogMessage instances,
when an msg is write-protected no changes can be applied. If a LogPipe
needs write access, it needs to acquire that by calling
log_msg_make_writable().

Write-protection is enabled on a LogMessage when it is assumed not to
change. Such an occassion is when it is put on a LogQueue of a destination
which processes messages asynchronously. A similar case is when the
db-parser() puts the message in its correllation state.

Before any writes can be applied, a call to log_msg_make_writable()
needs to be made, which creates a clone if the current LogMessage
is write-protected, or leaves the message be if it isn't.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>