debian: Mark the systemd symlinks as conffiles
The symlinks under /etc/systemd are meant to be user-configurable,
that's why they are in /etc, so mark them as conffiles, so upgrading
syslog-ng-core will do the right thing, and preserve their state.

dh_installdeb does not do this for us, as that only marks files as
conffiles, not symlinks.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>