json-parser(): added marker() option
This patch implements a marker() option for json-parser(). Marker is
a string prefix that can be used to identify a message to be parsed. If
the marker is set and the incoming message doesn't start with it,
parsing is not performed and the message is dropped.

If the input indeed matches the marker, json parsing commences with the
string _after_ marker. White space between the marker and the json
payload is ignored.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>