autogen.sh: Create configure.gnu ourselves in submodules
To make it easier to use upstream git repos as submodules, which do
not have configure.gnu, create those from within autogen.sh.

This way, it becomes much easier to switch to an upstream repo, and
still continue reaping the benefits the configure.gnu brings us (such
as forcing a static linkage).

Signed-off-by: Gergely Nagy <algernon@balabit.hu>