debian: Fix build-indep and multiple builds in a row
Instead of changing directory to $SRCDIR and running dh_autoreconf
there, stay in the current dir, and ask autoreconf to cd itself. This
results in the dh_autoreconf state files being put to the right place,
and they'll clean up nicely.

The above fixes the multiple builds in a row case, but since it also
removes the test that prevented dh_autoreconf from running twice,
broke the build-indep target, which tried to run the whole lot of
things again.

Since our single arch:all package does not need anything built at all,
the build-indep target should be a no-op instead.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>