debian: Fix arch-indep-only build
When building an arch-indep package, we were running into the
dh_auto_install override, which tried to install ivykis headers
too. Since we do not need that (and they're not available anyway) when
doing an arch-indep only build, this is a mistake.

Instead, only override dh_auto_install for the -arch case, and leave
-indep alone.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>