dnscache: Fix a memory corruption when destroying the DNS cache
The DNS cache gets destroyed every time a worker thread quits, which
is fine, as most of the dns cache variables are thread local.

However, dns_cache_hosts is not, it's a global static, and
dns_cache_destroy() was freeing that aswell.

The solution is to not free dns_cache_hosts when a worker stops, but
do so when syslog-ng stops. This patch introduces dns_cache_deinit()
which does just that, and removes the same thing from
dns_cache_destroy(), which now only touches thread-local variables.

Reported-by: EgonB <egon@local.ee>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>