afprog: Completely rebuild the writer on SIGHUP
When syslog-ng receives a SIGHUP, it restarts the program()
destinations, by sending them a SIGTERM, and starting a new
one. However, programs that ignore signals, and watch only their
stdin, remain running in this case, which is unfortunate.

To remedy the situation, and because we're already restarting the
program destinations, unref the writer, thus triggering the closing of
the file descriptor too. This way, we'll rebuild self->writer from
scratch on each HUP, but due to restarting the program anyway, this is
no big deal.

Reported-by: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>