logproto: Move LogProtoStatus into LogProto itself
Move the LogProtoStatus from LogProtoBufferedServer into LogProto
itself, so it can be reset or acted upon on a higher level. This
allows us to move the error state handling to logo_proto_fetch, and it
also allows us to clear any error state on a higher level.

With these changes, syslog-ng will correctly handle character encoding
issues on dgram sources: drop and the line with error, but clear the
error state and continue afterwards.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>