patterndb: Fix a NULL dereference in certain cases
In patterndb, all of the context-* properties are optional, yet, in
certain cases where they were only partially set (eg, context-id and
context-time set, but context-scope not), running an action ended up
with syslog-ng trying to dereference a NULL pointer.

This patch corrects that mistake, by teaching pdb_rule_run_actions()
to only dereference the context if it is non-NULL. The bulk of the
function handled the case well, just not the initial condition that
governs the core of the function.

Reported-by: Imre Lazar <imre@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>