lib/: Use upstream ivykis
This patch converts syslog-ng to use upstream ivykis,
unmodified. Apart from the boring struct and type renaming stuff,
there is one major change: iv_fd_pollable() is gone, we have
iv_fd_register_try().

As such, the logic in log_reader_start_watches() and
log_writer_start_watches() was updated to use the upstream API
instead.

The control socket code was also updated to use iv_fd_register_try(),
because we want synchronous registration in that case, and
iv_fd_register_try() provides that for us.

Furthermore, we now use iv_set_fatal_msg_handler() to print ivykis
errors to stderr, instead of syslog (this is functionally equivalent
to what the balabit-patched ivykis did).

The included ivykis is also updated to 0.30 (+ two small patches), and
the minimum ivykis requirement when compiling with a system ivykis has
been increased to 0.30 aswell, and the linking magic updated to use
the combined libivykis, and not search for ivykis-modules anymore.

And since the git submodule now points to upstream, and the
configure.gnu for ivykis is generated by our own autogen.sh, we need
to get that into the dist tarball. This is accomplished by including
it in lib/Makefile.am's EXTRA_DIST.

Many thanks to Lennert Buytenhek <buytenh@wantstofly.org> and Balazs
Scheidler <bazsi@balabit.hu> for their invaluable help.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>