afsql: Create proper indexes for Oracle DBs.
When creating an index for an Oracle DB, the field name was enclosed
in apostrophes, which made Oracle very unhappy: it did not associate
the index with the intended field.

With other databases, the field name is not quoted in the CREATE INDEX
statement, and for Oracle, it shouldn't be, either.

This patch corrects the mistake, and removes the apostrophes in the
Oracle case too.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>