LogProtoBufferedServer: fixed a SockAddr caching problem that shouldn't be user-visible
The log_proto_buffered_server_fetch() function improperly NULLed the
cached GSockAddr instance every time it attempted to read. However that
forthcoming read may return EAGAIN, which causes the sockaddr to remain
NULL.

In case an EOF is read the next time, log_proto_fetch() will return NULL as
the originating sockaddr for the partial message still in the buffer.

This bug shouldn't bite anyone as the cached sockaddr is only used for
udp() sources, but which in turn don't have partial messages nor EOFs.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>