LogProtoFramedServer: change dynamic buffer allocation
Until now the buffer of LogProtoFramedServer class implementing
RFC5424 style framing was sized in accordance to the maximum frame length
processed by the class. More exactly it was sized as "max_frame_len + 10".

This could mean that if most of the messages are about the same size, that
the buffer is limited to hold one single message. This is certainly not
the best for performance, so be a bit more liberal when allocating this
buffer size, it is scaled to hold 16 messages, but at the same time limited
to 6*log_msg_size().

This is 48k buffer space for the default log_msg_size() value of 8k and
matches exactly with the buffer size used by LogProtoTextServer.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>