LogQueue: fixed a memory leak in log_queue_fifo_push_head()
log_queue_fifo_push_head() is used by various drivers to push one item in
the front of the queue, but unlike log_queue_fifo_push_tail(), this one
didn't consume the passed reference.

This patch fixes this API breakage to behave the same as push_tail(),
and I also reviewed all the call-sites. Except for LogWriter that
was fixed just recently all others misused this part of the API. This
by itself would explain the change.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>